+++ tools/level_editor/properties_view.vala
@@ -47,7 +47,8 @@ public class UnitView : PropertyGrid
 			OBJECT_TYPE_SCRIPT,
 			OBJECT_TYPE_ANIMATION_STATE_MACHINE,
 			OBJECT_TYPE_FOG,
-			OBJECT_TYPE_GLOBAL_LIGHTING
+			OBJECT_TYPE_GLOBAL_LIGHTING,
+			OBJECT_TYPE_BLOOM,
 		};
 
 		// Construct 'add components' button.
@@ -167,6 +168,7 @@ public class PropertiesView : Gtk.Box
 		register_object_type(OBJECT_TYPE_SOUND_SOURCE,            0, null, UnitView.component_menu);
 		register_object_type(OBJECT_TYPE_FOG,                     0, null, UnitView.component_menu);
 		register_object_type(OBJECT_TYPE_GLOBAL_LIGHTING,         0, null, UnitView.component_menu);
+		register_object_type(OBJECT_TYPE_BLOOM,                   0, null, UnitView.component_menu);
 	}
 
 	private void register_object_type(string object_type, int position, PropertyGrid? cv = null, ContextMenu? context_menu = null)